dom/imptests/html/dom/nodes/test_DOMImplementation-createDocument.html

<!doctype html>
<meta charset=utf-8>
<title>DOMImplementation.createDocument(namespace, qualifiedName, doctype)</title>
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-domimplementation-createdocument">
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-document-createelementns">
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-node-nodetype">
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-document-documentelement">
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-document-doctype">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var tests = [
    /* Arrays with four elements:
     *   the namespace argument
     *   the qualifiedName argument
     *   the doctype argument
     *   the expected exception, or null if none
     */
    [null, null, false, new TypeError()],
    [null, null, null, null],
    [null, undefined, null, null],
    [null, "", null, null],
    [null, "foo", null, null],
    [null, "1foo", null, "INVALID_CHARACTER_ERR"],
    [null, "f1oo", null, null],
    [null, "foo1", null, null],
    [null, ":foo", null, "NAMESPACE_ERR"],
    [null, "f:oo", null, "NAMESPACE_ERR"],
    [null, "foo:", null, "NAMESPACE_ERR"],
    [null, "xml", null, null],
    [null, "xmlns", null, "NAMESPACE_ERR"],
    [null, "xmlfoo", null, null],
    [null, "xml:foo", null, "NAMESPACE_ERR"],
    [null, "xmlns:foo", null, "NAMESPACE_ERR"],
    [null, "xmlfoo:bar", null, "NAMESPACE_ERR"],
    [undefined, null, undefined, null],
    [undefined, undefined, undefined, null],
    [undefined, "", undefined, null],
    [undefined, "foo", undefined, null],
    [undefined, "1foo", undefined, "INVALID_CHARACTER_ERR"],
    [undefined, "f1oo", undefined, null],
    [undefined, "foo1", undefined, null],
    [undefined, ":foo", undefined, "NAMESPACE_ERR"],
    [undefined, "f:oo", undefined, "NAMESPACE_ERR"],
    [undefined, "foo:", undefined, "NAMESPACE_ERR"],
    [undefined, "xml", undefined, null],
    [undefined, "xmlns", undefined, "NAMESPACE_ERR"],
    [undefined, "xmlfoo", undefined, null],
    [undefined, "xml:foo", undefined, "NAMESPACE_ERR"],
    [undefined, "xmlns:foo", undefined, "NAMESPACE_ERR"],
    [undefined, "xmlfoo:bar", undefined, "NAMESPACE_ERR"],
    ["http://example.com/", null, null, null],
    ["http://example.com/", "", null, null],
    ["http://example.com/", "foo", null, null],
    ["http://example.com/", "1foo", null, "INVALID_CHARACTER_ERR"],
    ["http://example.com/", "f1oo", null, null],
    ["http://example.com/", "foo1", null, null],
    ["http://example.com/", ":foo", null, "NAMESPACE_ERR"],
    ["http://example.com/", "f:oo", null, null],
    ["http://example.com/", "foo:", null, "NAMESPACE_ERR"],
    ["http://example.com/", "xml", null, null],
    ["http://example.com/", "xmlns", null, "NAMESPACE_ERR"],
    ["http://example.com/", "xmlfoo", null, null],
    ["http://example.com/", "xml:foo", null, "NAMESPACE_ERR"],
    ["http://example.com/", "xmlns:foo", null, "NAMESPACE_ERR"],
    ["http://example.com/", "xmlfoo:bar", null, null],
    ["/", null, null, null],
    ["/", "", null, null],
    ["/", "foo", null, null],
    ["/", "1foo", null, "INVALID_CHARACTER_ERR"],
    ["/", "f1oo", null, null],
    ["/", "foo1", null, null],
    ["/", ":foo", null, "NAMESPACE_ERR"],
    ["/", "f:oo", null, null],
    ["/", "foo:", null, "NAMESPACE_ERR"],
    ["/", "xml", null, null],
    ["/", "xmlns", null, "NAMESPACE_ERR"],
    ["/", "xmlfoo", null, null],
    ["/", "xml:foo", null, "NAMESPACE_ERR"],
    ["/", "xmlns:foo", null, "NAMESPACE_ERR"],
    ["/", "xmlfoo:bar", null, null],
    ["http://www.w3.org/XML/1998/namespace", null, null, null],
    ["http://www.w3.org/XML/1998/namespace", "", null, null],
    ["http://www.w3.org/XML/1998/namespace", "foo", null, null],
    ["http://www.w3.org/XML/1998/namespace", "1foo", null, "INVALID_CHARACTER_ERR"],
    ["http://www.w3.org/XML/1998/namespace", "f1oo", null, null],
    ["http://www.w3.org/XML/1998/namespace", "foo1", null, null],
    ["http://www.w3.org/XML/1998/namespace", ":foo", null, "NAMESPACE_ERR"],
    ["http://www.w3.org/XML/1998/namespace", "f:oo", null, null],
    ["http://www.w3.org/XML/1998/namespace", "foo:", null, "NAMESPACE_ERR"],
    ["http://www.w3.org/XML/1998/namespace", "xml", null, null],
    ["http://www.w3.org/XML/1998/namespace", "xmlns", null, "NAMESPACE_ERR"],
    ["http://www.w3.org/XML/1998/namespace", "xmlfoo", null, null],
    ["http://www.w3.org/XML/1998/namespace", "xml:foo", null, null],
    ["http://www.w3.org/XML/1998/namespace", "xmlns:foo", null, "NAMESPACE_ERR"],
    ["http://www.w3.org/XML/1998/namespace", "xmlfoo:bar", null, null],
    ["http://www.w3.org/2000/xmlns/", null, null, null],
    ["http://www.w3.org/2000/xmlns/", "", null, null],
    ["http://www.w3.org/2000/xmlns/", "foo", null, "NAMESPACE_ERR"],
    ["http://www.w3.org/2000/xmlns/", "1foo", null, "INVALID_CHARACTER_ERR"],
    ["http://www.w3.org/2000/xmlns/", "f1oo", null, "NAMESPACE_ERR"],
    ["http://www.w3.org/2000/xmlns/", "foo1", null, "NAMESPACE_ERR"],
    ["http://www.w3.org/2000/xmlns/", ":foo", null, "NAMESPACE_ERR"],
    ["http://www.w3.org/2000/xmlns/", "f:oo", null, "NAMESPACE_ERR"],
    ["http://www.w3.org/2000/xmlns/", "foo:", null, "NAMESPACE_ERR"],
    ["http://www.w3.org/2000/xmlns/", "xml", null, "NAMESPACE_ERR"],
    ["http://www.w3.org/2000/xmlns/", "xmlns", null, null],
    ["http://www.w3.org/2000/xmlns/", "xmlfoo", null, "NAMESPACE_ERR"],
    ["http://www.w3.org/2000/xmlns/", "xml:foo", null, "NAMESPACE_ERR"],
    ["http://www.w3.org/2000/xmlns/", "xmlns:foo", null, null],
    ["http://www.w3.org/2000/xmlns/", "xmlfoo:bar", null, "NAMESPACE_ERR"],
    ["foo:", null, null, null],
    ["foo:", "", null, null],
    ["foo:", "foo", null, null],
    ["foo:", "1foo", null, "INVALID_CHARACTER_ERR"],
    ["foo:", "f1oo", null, null],
    ["foo:", "foo1", null, null],
    ["foo:", ":foo", null, "NAMESPACE_ERR"],
    ["foo:", "f:oo", null, null],
    ["foo:", "foo:", null, "NAMESPACE_ERR"],
    ["foo:", "xml", null, null],
    ["foo:", "xmlns", null, "NAMESPACE_ERR"],
    ["foo:", "xmlfoo", null, null],
    ["foo:", "xml:foo", null, "NAMESPACE_ERR"],
    ["foo:", "xmlns:foo", null, "NAMESPACE_ERR"],
    ["foo:", "xmlfoo:bar", null, null],
    [null, null, document.implementation.createDocumentType("foo", "", ""), null],
    [null, null, document.doctype, null], // This causes a horrible WebKit bug (now fixed in trunk).
    [null, null, function() {
        var foo = document.implementation.createDocumentType("foo", "", "");
        document.implementation.createDocument(null, null, foo);
        return foo;
     }(), null], // DOCTYPE already associated with a document.
    [null, null, function() {
        var bar = document.implementation.createDocument(null, null, null);
        return bar.implementation.createDocumentType("bar", "", "");
     }(), null], // DOCTYPE created by a different implementation.
    [null, null, function() {
        var bar = document.implementation.createDocument(null, null, null);
        var magic = bar.implementation.createDocumentType("bar", "", "");
        bar.implementation.createDocument(null, null, magic);
        return magic;
     }(), null], // DOCTYPE created by a different implementation and already associated with a document.
    [null, "foo", document.implementation.createDocumentType("foo", "", ""), null],
    ["foo", null, document.implementation.createDocumentType("foo", "", ""), null],
    ["foo", "bar", document.implementation.createDocumentType("foo", "", ""), null],
  ]

  tests.forEach(function(t, i) {
    test(function() {
      var namespace = t[0], qualifiedName = t[1], doctype = t[2], expected = t[3]
      if (expected != null) {
        assert_throws(expected, function() { document.implementation.createDocument(namespace, qualifiedName, doctype) })
      } else {
        var doc = document.implementation.createDocument(namespace, qualifiedName, doctype)
        assert_equals(doc.nodeType, Node.DOCUMENT_NODE)
        assert_equals(doc.nodeType, doc.DOCUMENT_NODE)
        var omitRootElement = qualifiedName === null || String(qualifiedName) === ""
        if (omitRootElement) {
          assert_equals(doc.documentElement, null)
        } else {
          var element = doc.documentElement
          assert_not_equals(element, null)
          assert_equals(element.nodeType, Node.ELEMENT_NODE)
          assert_equals(element.ownerDocument, doc)
          var qualified = String(qualifiedName), names = []
          if (qualified.indexOf(":") >= 0) {
            names = qualified.split(":", 2)
          } else {
            names = [null, qualified]
          }
          assert_equals(element.prefix, names[0])
          assert_equals(element.localName, names[1])
          assert_equals(element.namespaceURI, namespace === undefined ? null : namespace)
        }
        if (!doctype) {
          assert_equals(doc.doctype, null)
        } else {
          assert_equals(doc.doctype, doctype)
          assert_equals(doc.doctype.ownerDocument, doc)
        }
        assert_equals(doc.childNodes.length, !omitRootElement + !!doctype)
      }
    }, "createDocument test " + i + ": " + t.map(function(el) { return format_value(el) }))
  })
})
</script>
